backend/requirements.txt
Flask==2.2.3
torch==2.2.0
transformers==4.26.0
diffusers==0.18.0
requests==2.28.1
Pillow==9.0.1
